Split headers and remove spurious files
